apps/purchase/backend/requirements.txt
flask
flask-sqlalchemy
flask-cors
pandas
python-dotenv
